+++ src/init.lua
@@ -30,7 +30,7 @@ return {
 
 	-- Navigation Actions
 	Actions = require(script.NavigationActions),
-	StackActions = require(script.StackActions),
+	StackActions = require(script.routers.StackActions),
 	SwitchActions = require(script.routers.SwitchActions),
 	BackBehavior = require(script.BackBehavior),