+++ src/routers/StackActions.lua
@@ -1,14 +1,16 @@
 -- upstream https://github.com/react-navigation/react-navigation/blob/62da341b672a83786b9c3a80c8a38f929964d7cc/packages/core/src/routers/StackActions.js
 
-local Packages = script.Parent.Parent
+local root = script.Parent.Parent
+local Packages = root.Parent
 local Cryo = require(Packages.Cryo)
-local NavigationSymbol = require(script.Parent.NavigationSymbol)
+local NavigationSymbol = require(root.NavigationSymbol)
 
 local POP_TOKEN = NavigationSymbol("POP")
 local POP_TO_TOP_TOKEN = NavigationSymbol("POP_TO_TOP")
 local PUSH_TOKEN = NavigationSymbol("PUSH")
 local RESET_TOKEN = NavigationSymbol("RESET")
 local REPLACE_TOKEN = NavigationSymbol("REPLACE")
+local COMPLETE_TRANSITION_TOKEN = NavigationSymbol("COMPLETE_TRANSITION")
 
 --[[
 	StackActions provides shared constants and methods to construct
@@ -22,6 +24,7 @@ local StackActions = {
 	Push = PUSH_TOKEN,
 	Reset = RESET_TOKEN,
 	Replace = REPLACE_TOKEN,
+	CompleteTransition = COMPLETE_TRANSITION_TOKEN,
 }
 
 -- deviation: we using this metatable to error when StackActions is indexed
@@ -61,6 +64,11 @@ function StackActions.replace(payload)
 	)
 end
 
--- deviation: `StackActions.CompleteTransition` is moved in NavigationActions
+function StackActions.completeTransition(payload)
+	return Cryo.Dictionary.join(
+		{ type = COMPLETE_TRANSITION_TOKEN, preserveFocus = true },
+		payload or {}
+	)
+end
 
 return StackActions